WRONG
Attempt to generalize over irrelevant empty cells

This does NOT work as is. It correctly works out the winner, but can be wrong about the value since it cannot predict how many (even sets) of irrelevant spaces will get filled.
So this can be used for the weak calcualtion or [-1,0], [0,1], [-1,1] windows but not in general